requirements.txt
flask-cors
flask
yagmail
python-dotenv
